+++ db/uninstall.php
@@ -15,18 +15,29 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Theme Boost Campus - Version file
+ * Theme Boost Campus - Uninstall file.
  *
  * @package    theme_boost_campus
- * @copyright  2017 Kathrin Osswald, Ulm University <kathrin.osswald@uni-ulm.de>
+ * @copyright  2020 Kathrin Osswald, Ulm University <kathrin.osswald@uni-ulm.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->component = 'theme_boost_campus';
-$plugin->version = 2020080500;
-$plugin->release = 'v3.8-r2';
-$plugin->requires = 2019111804;
-$plugin->maturity = MATURITY_STABLE;
-$plugin->dependencies = array('theme_boost' => 2019111800);
+/**
+ * Plugin uninstall steps.
+ */
+function xmldb_theme_boost_campus_uninstall() {
+    global $DB;
+
+    // The plugin uninstall process in Moodle core will take care of removing the plugin configuration, but not of removing the
+    // user preferences which we have set for the users. We have to remove them ourselves.
+    // We remove them directly from the DB table and don't use unset_user_preference() as the cache is cleared anyway directly
+    // after the plugin has been uninstalled.
+
+    $like = $DB->sql_like('name', '?', true, true, false, '|');
+    $params = array($DB->sql_like_escape('theme_boost_campus-', '|') . '%');
+    $DB->delete_records_select('user_preferences', $like, $params);
+
+    return true;
+}