+++ layout/columns2.php
@@ -29,6 +29,9 @@ global $PAGE;
 // MODIFICATION END.
 
 user_preference_allow_ajax_update('drawer-open-nav', PARAM_ALPHA);
+// MODIFICATION START: Allow own user preference to be set via Javascript.
+user_preference_allow_ajax_update('theme_boost_campus_infobanner_dismissed', PARAM_BOOL);
+// MODIFICATION END.
 require_once($CFG->libdir . '/behat/lib.php');
 // MODIFICATION Start: Require own locallib.php.
 require_once($CFG->dirroot . '/theme/boost_campus/locallib.php');
@@ -76,6 +79,11 @@ if (get_config('theme_boost_campus', 'darknavbar') == 'yes') {
 $navdrawerfullwidth = get_config('theme_boost_campus', 'navdrawerfullwidth');
 // MODIFICATION END.
 
+// MODIFICATION START: Set these variables in any case as it's needed in the columns2.mustache file.
+$perpinfobannershowonselectedpage = false;
+$timedinfobannershowonselectedpage = false;
+// MODIFICATION END.
+
 $templatecontext = [
     'sitename' => format_string($SITE->shortname, true, ['context' => context_course::instance(SITEID), "escape" => false]),
     'output' => $OUTPUT,
@@ -85,12 +93,61 @@ $templatecontext = [
     'navdraweropen' => $navdraweropen,
     'regionmainsettingsmenu' => $regionmainsettingsmenu,
     'hasregionmainsettingsmenu' => !empty($regionmainsettingsmenu),
-    // MODIFICATION START: Add Boost Campus realated values to the template context.
+    // MODIFICATION START: Add Boost Campus related values to the template context.
     'catchshortcuts' => json_encode($catchshortcuts),
     'navdrawerfullwidth' => $navdrawerfullwidth,
-    'darknavbar' => $darknavbar
+    'darknavbar' => $darknavbar,
+    'perpinfobannershowonselectedpage' => $perpinfobannershowonselectedpage,
+    'timedinfobannershowonselectedpage' => $timedinfobannershowonselectedpage
     // MODIFICATION END.
 ];
+
+// MODIFICATION START: Settings for perpetual information banner.
+$perpibenable = get_config('theme_boost_campus', 'perpibenable');
+
+if ($perpibenable) {
+    $perpibcontent = format_text(get_config('theme_boost_campus', 'perpibcontent'), FORMAT_HTML);
+    // Result of multiselect is a string divided by a comma, so exploding into an array.
+    $perpibshowonpages = explode(",", get_config('theme_boost_campus', 'perpibshowonpages'));
+    $perpibcss = get_config('theme_boost_campus', 'perpibcss');
+    $perpibdismiss = get_config('theme_boost_campus', 'perpibdismiss');
+    $perbibconfirmdialogue = get_config('theme_boost_campus', 'perpibconfirm');
+    $perbibuserprefdialdismissed = get_user_preferences('theme_boost_campus_infobanner_dismissed');
+
+    $perpinfobannershowonselectedpage = theme_boost_campus_show_banner_on_selected_page($perpibshowonpages,
+            $perpibcontent, $PAGE->pagelayout, $perbibuserprefdialdismissed);
+
+    // Add the variables to the templatecontext array.
+    $templatecontext['perpibcontent'] = $perpibcontent;
+    $templatecontext['perpibcss'] = $perpibcss;
+    $templatecontext['perpibdismiss'] = $perpibdismiss;
+    $templatecontext['perpinfobannershowonselectedpage'] = $perpinfobannershowonselectedpage;
+    $templatecontext['perbibconfirmdialogue'] = $perbibconfirmdialogue;
+}
+// MODIFICATION END.
+
+// MODIFICATION START: Settings for time controlled information banner.
+$timedibenable = get_config('theme_boost_campus', 'timedibenable');
+
+if ($timedibenable) {
+    $timedibcontent = format_text(get_config('theme_boost_campus', 'timedibcontent'), FORMAT_HTML);
+    // Result of multiselect is a string divided by a comma, so exploding into an array.
+    $timedibshowonpages = explode(",", get_config('theme_boost_campus', 'timedibshowonpages'));
+    $timedibcss = get_config('theme_boost_campus', 'timedibcss');
+    $timedibstartsetting = get_config('theme_boost_campus', 'timedibstart');
+    $timedibendsetting = get_config('theme_boost_campus', 'timedibend');
+    // Get the current server time.
+    $now = (new DateTime("now", core_date::get_server_timezone_object()))->getTimestamp();
+
+    $timedinfobannershowonselectedpage = theme_boost_campus_show_timed_banner_on_selected_page($now, $timedibshowonpages,
+            $timedibcontent, $timedibstartsetting, $timedibendsetting, $PAGE->pagelayout);
+
+    // Add the variables to the templatecontext array.
+    $templatecontext['timedibcontent'] = $timedibcontent;
+    $templatecontext['timedibcss'] = $timedibcss;
+    $templatecontext['timedinfobannershowonselectedpage'] = $timedinfobannershowonselectedpage;
+}
+// MODIFICATION END.
 
 $nav = $PAGE->flatnav;
 // MODIDFICATION START.